Fixed issue which led to broken fingerprints.
When cropping, Papercrop would update the cropped files, but not update
anything in the database aside from the updated_at column. This was
efficient, but led to problems in certain circumstances. Specifically,
if you had defined a style for :original (in order to constrain the size
of the stored original image), using Papercrop could cause a
modification of the original image due to it running through its
processors again. This would cause its fingerprint to change, but the
fingerprint would not get persisted in the database. If you were using
the fingerprint in your filenames, all the filenames would get changed,
but Paperclip couldn't build working paths to the files anymore because
the fingerprint in the database was still the old fingerprint!